move LogConfig
