+++ common-lib/src/main/scala/com/gu/mediaservice/lib/logging/LogConfig.scala
@@ -1,4 +1,4 @@
-package lib
+package com.gu.mediaservice.lib.logging
 
 import ch.qos.logback.classic.spi.ILoggingEvent
 import ch.qos.logback.classic.{LoggerContext, Logger => LogbackLogger}
@@ -9,20 +9,17 @@ import org.slf4j.{LoggerFactory, Logger => SLFLogger}
 import play.api.ApplicationLoader.Context
 import play.api.LoggerConfigurator
 import play.api.libs.json._
-
-import scala.util.Try
 import scalaz.syntax.id._
+import scala.util.Try
 
 
 object LogConfig {
 
-  val rootLogger = LoggerFactory.getLogger(SLFLogger.ROOT_LOGGER_NAME).asInstanceOf[LogbackLogger]
+  val rootLogger: LogbackLogger = LoggerFactory.getLogger(SLFLogger.ROOT_LOGGER_NAME).asInstanceOf[LogbackLogger]
 
   case class KinesisAppenderConfig(stream: String, region: String, roleArn: String, bufferSize: Int)
 
-
-
-  def makeCustomFields(config: CommonConfig) = {
+  def makeCustomFields(config: CommonConfig): String = {
     Json.toJson(Map(
       "stack" -> config.stackName,
       "stage" -> config.stage.toUpperCase,
@@ -31,10 +28,10 @@ object LogConfig {
     )).toString()
   }
 
-  def makeLayout(customFields: String) = (new LogstashLayout()) <| (_.setCustomFields(customFields))
+  def makeLayout(customFields: String) = new LogstashLayout() <| (_.setCustomFields(customFields))
 
   def makeKinesisAppender(layout: LogstashLayout, context: LoggerContext, appenderConfig: KinesisAppenderConfig) =
-    (new KinesisAppender[ILoggingEvent]()) <| { a =>
+    new KinesisAppender[ILoggingEvent]() <| { a =>
       a.setStreamName(appenderConfig.stream)
       a.setRegion(appenderConfig.region)
       a.setRoleToAssumeArn(appenderConfig.roleArn)
@@ -47,30 +44,33 @@ object LogConfig {
       a.start()
   }
 
-  def initKinesisLogging(config: CommonConfig): Unit = config.stage match {
-    case "DEV" =>  rootLogger.info("Logging disabled in DEV")
-    case _  => Try {
-      rootLogger.info("LogConfig initializing")
-      rootLogger.info("Configuring Logback")
+  def initKinesisLogging(config: CommonConfig): Unit = {
+    if (config.isDev) {
+      rootLogger.info("Kinesis logging disabled in DEV")
+    } else {
+      Try {
+        rootLogger.info("LogConfig initializing")
+        rootLogger.info("Configuring Logback")
 
-      val customFields = makeCustomFields(config)
-      val context      = rootLogger.getLoggerContext
-      val layout       = makeLayout(customFields)
-      val bufferSize   = 1000
+        val customFields = makeCustomFields(config)
+        val context      = rootLogger.getLoggerContext
+        val layout       = makeLayout(customFields)
+        val bufferSize   = 1000
 
-      val appender     = makeKinesisAppender(layout, context,
-        KinesisAppenderConfig(
-          config.properties("logger.kinesis.stream"),
-          config.properties("logger.kinesis.region"),
-          config.properties("logger.kinesis.roleArn"),
-          bufferSize
+        val appender     = makeKinesisAppender(layout, context,
+          KinesisAppenderConfig(
+            config.properties("logger.kinesis.stream"),
+            config.properties("logger.kinesis.region"),
+            config.properties("logger.kinesis.roleArn"),
+            bufferSize
+          )
         )
-      )
 
-      rootLogger.addAppender(appender)
-      rootLogger.info("Configured Logback")
-    } recover {
-      case e => rootLogger.error("LogConfig Failed!", e)
+        rootLogger.addAppender(appender)
+        rootLogger.info("Configured Logback")
+      } recover {
+        case e => rootLogger.error("LogConfig Failed!", e)
+      }
     }
   }
